添加许愿墙

+++ china_service/src/main/java/blct/china/service/sticker/entity/StickerExample.java
@@ -305,7 +305,7 @@ public class StickerExample {
             return (Criteria) this;
         }
 
-        public Criteria andValidEqualTo(Byte value) {
+        public Criteria andValidEqualTo(Integer value) {
             addCriterion("valid =", value, "valid");
             return (Criteria) this;
         }
